Move carousel data retrieval operations out of generic carousel
- The CardCarouselGeneric component was handling the UI and data retrieval
- Now CardCarouselGeneric handles UI while CardCarouselHydroshareFeatured handles the data retrieval
- Improves design through separation of concerns
- Updated the Apps page to use the new component
- Reduces repeated code when adding the featured carousel to other pages
